+++ Metalama.Framework.Engine/Linking/IntermediateSymbolSemanticEqualityComparer.cs
@@ -59,10 +59,10 @@ namespace Metalama.Framework.Engine.Linking
 
         public int GetHashCode( IntermediateSymbolSemantic<TSymbol> x )
         {
-            // PERF: Cast enum to int otherwise it will be boxed within the method.
+            // PERF: Cast enum to byte otherwise it will be boxed within the method.
             return HashCode.Combine(
                 this._symbolComparer.GetHashCode( x.Symbol ),
-                x.Kind );
+                (byte)x.Kind );
         }
     }
 }
